pyproject.toml
[project]
name = "attendance-control-backend"
version = "0.1.0"
description = "The backend of the attendance control system"
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "email-validator<3.0.0,>=2.2.0",
    "fastapi<1.0.0,>=0.115.12",
    "pydantic-settings<3.0.0,>=2.9.1",
    "sqlmodel<1.0.0,>=0.0.24",
    "alembic<2.0.0,>=1.16.2",
    "fastapi-cli<1.0.0,>=0.0.7",
    "psycopg[binary]<4.0.0,>=3.2.9",
    "bcrypt>=4.3.0",
    "python-multipart>=0.0.20",
    "pyjwt>=2.10.1",
]

[dependency-groups]
dev = [
    "httpx<1.0.0,>=0.28.1",
    "mypy<2.0.0,>=1.16.0",
    "pre-commit<5.0.0,>=4.2.0",
    "pytest<9.0.0,>=8.4.0",
    "ruff<1.0.0,>=0.11.13",
    "types-passlib<2.0.0,>=1.7.7.20250602",
]

[tool.pytest.ini_options]
testpaths = ["app/tests"]

[tool.ruff]
target-version = "py312"
exclude = ["app/alembic"]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG001", # unused arguments in functions
]
ignore = [
    "B904",  # Allow raising exceptions without from e, for HTTPException
]
